+++ src/main/kotlin/github/ablandel/anotherkaomoji/repository/KaomojiRepository.kt
@@ -4,6 +4,7 @@ import github.ablandel.anotherkaomoji.entity.Kaomoji
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
+import java.util.*
 
 @Repository
 interface KaomojiRepository : CrudRepository<Kaomoji, Long> {
@@ -15,4 +16,7 @@ interface KaomojiRepository : CrudRepository<Kaomoji, Long> {
 
     @Query("SELECT k FROM kaomoji k WHERE LOWER(k.key) = LOWER(:key) OR k.emoticon = :emoticon")
     fun findByKeyOrEmoticon(key: String, emoticon: String): List<Kaomoji>
+
+    fun findByKeyIgnoreCase(key: String): Optional<Kaomoji>
+    fun findByEmoticon(emoticon: String): Optional<Kaomoji>
 }